Fixed constant calls to SDK due to Full Name failed request
(cherry picked from commit 409530c34bc2a55670c71b3efc032f3b90c06ff8)

+++ src/MEGASync/gui/UserAttributesRequests/Avatar.cpp
@@ -61,19 +61,22 @@ void Avatar::onRequestFinish(mega::MegaApi*, mega::MegaRequest* incoming_request
             }
             if (mFullName)
             {
-                disconnect(mFullName.get(), &FullName::attributeReady, this, &Avatar::onFullNameAttributeReady);
+                disconnect(mFullName.get(), &FullName::fullNameReady, this, &Avatar::onFullNameAttributeReady);
             }
         }
         else
         {
-            QFile::remove(mIconPath);
-            mIconPath.clear();
+            if(!mIconPath.isEmpty())
+            {
+                QFile::remove(mIconPath);
+                mIconPath.clear();
+            }
             mIcon.clear();
             if (!mFullName)
             {
                 mFullName = FullName::requestFullName(getEmail().toUtf8().constData());
             }
-            connect(mFullName.get(), &FullName::attributeReady, this, &Avatar::onFullNameAttributeReady);
+            connect(mFullName.get(), &FullName::fullNameReady, this, &Avatar::onFullNameAttributeReady, Qt::UniqueConnection);
             onFullNameAttributeReady();
         }
 
@@ -93,9 +96,7 @@ AttributeRequest::RequestInfo Avatar::fillRequestInfo()
         MegaSyncApp->getMegaApi()->getUserAvatar(getEmail().toUtf8().constData(),
                                                  Utilities::getAvatarPath(getEmail()).toUtf8().constData());
     };
-    QSharedPointer<ParamInfo> avatarParamInfo(new ParamInfo(avatarRequestFunc, QList<int>()
-                                                            << mega::MegaError::API_OK
-                                                            << mega::MegaError::API_ENOENT));
+    QSharedPointer<ParamInfo> avatarParamInfo(new ParamInfo(avatarRequestFunc));
     ParamInfoMap paramInfo({{mega::MegaApi::USER_ATTR_AVATAR, avatarParamInfo}});
     RequestInfo ret(paramInfo, QMap<int, int>({{mega::MegaUser::CHANGE_TYPE_AVATAR,
                                                 mega::MegaApi::USER_ATTR_AVATAR}}));
